requirements.txt
Flask==2.0.1
PyPDF2==1.26.0
pytesseract==0.3.8
Pillow==8.2.0
spacy==3.0.6
